backend/requirements-auth.txt
"""
djongo
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-cors-headers
requests
"""
